refactor(generator): split router components to files
